issn only for serial publications
modified map_omid_openalex_ids() in order to consider only ISSNs if the entity type is in ['journal', 'series', 'book series'], i.e. is a serial publication. other entity types are discarded (not considered for the mapping, though present in the initial tables). This change has been made in order to avoid multi-mapping an OMID to OpenAlex IDs in situations where the same (serial publication) entity is both represented as a Work (with a DOI) and a Source (with an ISSN).