make tables from IDs in OC Meta venue field
add support for venue field in th function creating the preliminary "reduced" tables from the original OC Meta tables. Now, when reading the original Meta tables, two output tables are created for each input table: one storing the IDs (and the type) of the entities represented by the whole row (i.e. the entity to which the IDs in the 'id' field of the original ro are associated with); the other table storing the IDs of the entity represented in the 'venue' field of each row.