bug fix
manage two different cases for strings of responsible agents, in get_ra_ids: since the publisher field only contain one entity, while author and editor fields can contain multiple entities, only the latter must be split by the specific separator ('; '), while semi-colon is a "normal" character in the string of the publisher (therefore the value of publisher must not be split by "; "). This bug did not produce errors in the output ID tables, because the content of square brackets (i.e. the IDs) has always been retained (the program only avoided processing the sub-string of the value of the publisher field corresponding to a part of the name). Nonetheless, more than 4.000 cases in which the publisher name contains a semi-colon have been signalled as format errors in the log files (while it is perfectly ok for a string of a publisher name to contain a semi-colon, since in the case of the publisher field this character is not a separator) .