cleaning and restructuring
The whole multi-mapped categorization process is now in a single file, mm_categ, and is organized in classes (plus a couple of standalone functions). the process can be called via CLI with a config file (just like the other intense  processes), of which an example is provided in mm_categ_config.yaml.