completed analysis_guide notebook
analysis_guide.ipynb can be used as a tutorial providing partial help with the analysis of results. Guides on how to run the processes (mapping process; multi-mapped BRs categorisation; unmapped BRs provenance analysis) are still missing. The notebook only helps to perform basic counts and visualisations on data that is already available